requirements.txt
pandas
scikit-learn
seaborn
matplotlib
tensorflow==2.12.0
pyarrow
nltk
tensorboard
scikeras
fastapi==0.78.0
uvicorn==0.18.3